Move message handling to protocol module
