+++ packages/portalnetwork/src/wire/index.ts
@@ -1,3 +1,1 @@
 export * from './types'
-export { ContentLookup } from './contentLookup'
-export { NodeLookup } from './nodeLookup'
